Refactor ops into standalone modules and unify API
