+++ fastmlx/op/__init__.py
@@ -1,5 +1,31 @@
-"""Operation classes."""
+"""Operation classes used within FastMLX."""
 
 from .op import Op
+from .expand_dims import ExpandDims
+from .minmax import Minmax
+from .normalize import Normalize
+from .pad_if_needed import PadIfNeeded
+from .random_crop import RandomCrop
+from .horizontal_flip import HorizontalFlip
+from .coarse_dropout import CoarseDropout
+from .onehot import Onehot
+from .sometimes import Sometimes
+from .cross_entropy import CrossEntropy
+from .model_op import ModelOp
+from .update_op import UpdateOp
 
-__all__ = ["Op"]
+__all__ = [
+    "Op",
+    "ExpandDims",
+    "Minmax",
+    "Normalize",
+    "PadIfNeeded",
+    "RandomCrop",
+    "HorizontalFlip",
+    "CoarseDropout",
+    "Onehot",
+    "Sometimes",
+    "CrossEntropy",
+    "ModelOp",
+    "UpdateOp",
+]
