--- fastmlx/op/numpyop/__init__.py
@@ -1,26 +1,0 @@
-"""Simple preprocessing ops implemented with MLX."""
-
-from .univariate import (
-    ExpandDims,
-    Minmax,
-    Normalize,
-    PadIfNeeded,
-    RandomCrop,
-    HorizontalFlip,
-    CoarseDropout,
-    Onehot,
-    Sometimes,
-)
-
-__all__ = [
-    "ExpandDims",
-    "Minmax",
-    "Normalize",
-    "PadIfNeeded",
-    "RandomCrop",
-    "HorizontalFlip",
-    "CoarseDropout",
-    "Onehot",
-    "Sometimes",
-]
-
