--- fastmlx/op/tensorop/__init__.py
@@ -1,6 +1,0 @@
-"""Tensor operations implemented using MLX."""
-
-from .loss import CrossEntropy
-from .model import ModelOp, UpdateOp
-
-__all__ = ["CrossEntropy", "ModelOp", "UpdateOp"]
